\name{rkafkajars}
\alias{rkafkajars}
\alias{rkafkajars}
\docType{package}
\title{
External Jars Required for Package 'rkafka'
}

\description{
The 'rkafkajars' package collects all the external jars required for the 'rkafka' package. These external jars are quite large in size (12MB) and have a slow release cycle. By separating the Java and R development, the storage footprint on CRAN is reduced. By using the 'rJava' package that links R and Java, we can use the excellent work already done by the folks and provide the functionality of using 'KAFKA' in R

}
\details{
\tabular{ll}{
Package: \tab rkafkajars\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-10\cr
License: \tab Apache License 2.0\cr
}
}
\author{
Shruti Gupta

Maintainer: Shruti Gupta<shruti.gupta@mu-sigma.com>

}


